Conflicts:
	Sources/Internal/Render/2D/Systems/RenderSystem2D.cpp
	Sources/Internal/Render/2D/Systems/RenderSystem2D.h
	Sources/Internal/Render/Texture.cpp

Former-commit-id: 80a4a243cd6cadfd42576b7c2306a0ba1774a524

+++ Sources/Internal/Render/Highlevel/Landscape.cpp
@@ -1203,7 +1203,7 @@ void Landscape::Draw(Camera * camera)
 	}
         
     BindMaterial(nearLodIndex, camera);
-    int32 count0 = lod0quads.size();
+    int32 count0 = static_cast<int32>(lod0quads.size());
     for(int32 i = 0; i < count0; ++i)
     {
         DrawQuad(lod0quads[i], 0);
@@ -1215,7 +1215,7 @@ void Landscape::Draw(Camera * camera)
 		BindMaterial(farLodIndex, camera);
 	}
 
-    int32 countNot0 = lodNot0quads.size();
+    int32 countNot0 = static_cast<int32>(lodNot0quads.size());
     for(int32 i = 0; i < countNot0; ++i)
     {
         DrawQuad(lodNot0quads[i], lodNot0quads[i]->data.lod);
@@ -1258,7 +1258,7 @@ void Landscape::Draw(Camera * camera)
 		{
 			BindMaterial(nearLodIndex, camera);
 		}
-        int32 count0 = lod0quads.size();
+        int32 count0 = static_cast<int32>(lod0quads.size());
         for(int32 i = 0; i < count0; ++i)
         {
             DrawQuad(lod0quads[i], 0);
@@ -1270,7 +1270,7 @@ void Landscape::Draw(Camera * camera)
 			BindMaterial(farLodIndex, camera);
 		}
         
-        int32 countNot0 = lodNot0quads.size();
+        int32 countNot0 = static_cast<int32>(lodNot0quads.size());
         for(int32 i = 0; i < countNot0; ++i)
         {
             DrawQuad(lodNot0quads[i], lodNot0quads[i]->data.lod);
