+++ Sources/Internal/Scene3D/Components/Waypoint/PathComponent.cpp
@@ -74,7 +74,7 @@ void PathComponent::Waypoint::AddEdge(PathComponent::Edge *edge)
 
 void PathComponent::Waypoint::RemoveEdge(PathComponent::Edge *edge)
 {
-    uint32 edgesCount = edges.size();
+    uint32 edgesCount = static_cast<uint32>(edges.size());
     for(uint32 e = 0; e < edgesCount; ++e)
     {
         if(edge == edges[e])
@@ -169,7 +169,7 @@ Component * PathComponent::Clone(Entity * toEntity)
     newComponent->SetName(name);
 	newComponent->SetEntity(toEntity);
 
-    const uint32 waypointCount = waypoints.size();
+    const uint32 waypointCount = static_cast<uint32>(waypoints.size());
     if(waypointCount)
     {
         newComponent->waypoints.resize(waypointCount);
@@ -187,7 +187,7 @@ Component * PathComponent::Clone(Entity * toEntity)
             newWaypoint->position = waypoint->position;
             newWaypoint->SetProperties(waypoint->GetProperties());
 
-            const uint32 edgesCount = waypoint->edges.size();
+            const uint32 edgesCount = static_cast<uint32>(waypoint->edges.size());
             for(uint32 e = 0; e < edgesCount; ++e)
             {
                 Edge *edge = waypoint->edges[e];
@@ -217,7 +217,7 @@ void PathComponent::Serialize(KeyedArchive *archive, SerializationContext *seria
     {
         archive->SetFastName("name", name);
         
-        const uint32 waypointCount = waypoints.size();
+        const uint32 waypointCount = static_cast<uint32>(waypoints.size());
         archive->SetUInt32("waypointCount", waypointCount);
 
         if(waypointCount)
@@ -235,7 +235,7 @@ void PathComponent::Serialize(KeyedArchive *archive, SerializationContext *seria
                     wpArchieve->SetArchive("properties", wp->GetProperties());
                 }
                 
-                const uint32 edgesCount = wp->edges.size();
+                const uint32 edgesCount = static_cast<uint32>(wp->edges.size());
                 wpArchieve->SetUInt32("edgesCount", edgesCount);
                 for(uint32 e = 0; e < edgesCount; ++e)
                 {
@@ -335,12 +335,12 @@ void PathComponent::AddPoint(DAVA::PathComponent::Waypoint *point)
 
 void PathComponent::RemovePoint(DAVA::PathComponent::Waypoint *point)
 {
-    uint32 waypointCount = waypoints.size();
+    uint32 waypointCount = static_cast<uint32>(waypoints.size());
     for(uint32 w = 0; w < waypointCount; ++w)
     {
         Waypoint *wp = waypoints[w];
         
-        uint32 edgesCount = wp->edges.size();
+        uint32 edgesCount = static_cast<uint32>(wp->edges.size());
         for(uint32 e = 0; e < edgesCount; ++e)
         {
             Edge *edge = wp->edges[e];
@@ -365,7 +365,7 @@ void PathComponent::RemovePoint(DAVA::PathComponent::Waypoint *point)
 
 PathComponent::Waypoint * PathComponent::GetWaypoint(const FastName & name)
 {
-    const uint32 waypointCount = waypoints.size();
+    const uint32 waypointCount = static_cast<uint32>(waypoints.size());
     for(uint32 w = 0; w < waypointCount; ++w)
     {
         Waypoint *wp = waypoints[w];
